requirements.txt
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.4
django-taggit==6.1.0
Markdown==3.7
psycopg==3.1.18
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
